refactor: move file watching and restart logic to separate module
