+++ packages/core/src/config.ts
@@ -2,7 +2,6 @@ import fs from 'node:fs';
 import { createRequire } from 'node:module';
 import { dirname, isAbsolute, join } from 'node:path';
 import { pathToFileURL } from 'node:url';
-import type { ChokidarOptions } from '../compiled/chokidar/index.js';
 import RspackChain from '../compiled/rspack-chain/index.js';
 import {
   ASSETS_DIST_DIR,
@@ -25,7 +24,6 @@ import {
 } from './constants';
 import {
   color,
-  debounce,
   findExists,
   getNodeEnv,
   isFileExists,
@@ -34,8 +32,6 @@ import {
 } from './helpers';
 import { logger } from './logger';
 import { mergeRsbuildConfig } from './mergeConfig';
-import { restartBuild, restartDevServer } from './server/restart';
-import { createChokidar } from './server/watchFiles.js';
 import type {
   InspectConfigOptions,
   InspectConfigResult,
@@ -369,42 +365,6 @@ const resolveConfigPath = (root: string, customConfig?: string) => {
 
   return null;
 };
-
-export async function watchFilesForRestart(
-  files: string[],
-  root: string,
-  isBuildWatch: boolean,
-  watchOptions?: ChokidarOptions,
-): Promise<void> {
-  if (!files.length) {
-    return;
-  }
-
-  const watcher = await createChokidar(files, root, {
-    // do not trigger add for initial files
-    ignoreInitial: true,
-    // If watching fails due to read permissions, the errors will be suppressed silently.
-    ignorePermissionErrors: true,
-    ...watchOptions,
-  });
-
-  const callback = debounce(
-    async (filePath) => {
-      await watcher.close();
-      if (isBuildWatch) {
-        await restartBuild({ filePath });
-      } else {
-        await restartDevServer({ filePath });
-      }
-    },
-    // set 300ms debounce to avoid restart frequently
-    300,
-  );
-
-  watcher.on('add', callback);
-  watcher.on('change', callback);
-  watcher.on('unlink', callback);
-}
 
 export type LoadConfigOptions = {
   /**
